+++ nostr-relay/nostr-relay-nip40/src/main/java/org/tbk/nostr/relay/nip40/interceptor/ExpiringEventHandlerInterceptor.java
@@ -26,8 +26,8 @@ public class ExpiringEventHandlerInterceptor implements RequestHandlerIntercepto
     }
 
     private void handleEvent(Event event) {
-        Nip40.getExpiration(event).ifPresent(expiresAt -> {
-            support.markExpiresAt(EventId.of(event), expiresAt).subscribe(unused -> {
+        Nip40.findExpiration(event).ifPresent(expiresAt -> {
+            support.markExpiresAt(EventId.of(event), expiresAt.instant()).subscribe(unused -> {
                 log.debug("Successfully marked event {} as expired.", event.getId());
             }, e -> {
                 log.warn("Error while marking event {} as expired: {}", event.getId(), e.getMessage());